requirements.txt
pandas
openpyxl
matplotlib
seaborn
fpdf2